docstrings_testing/news_recommender/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.0
bcrypt==4.0.1